refactor(phase-1): extract WizardState into models/wizard_state.py
Move 53-field WizardState dataclass out of app.py into
osx_proxmox_next/models/wizard_state.py with a clean __init__.py.
Remove unused fetch_vm_info, SmbiosIdentity imports from app.py.
Fix pre-existing unused CommandResult import in preflight.py.
