+++ src/osx_proxmox_next/models/__init__.py
@@ -1,0 +1,3 @@
+from .wizard_state import WizardState
+
+__all__ = ["WizardState"]
